Fixed alot of stuff
Bookmarks are mostly functional. You can add bookmarks via right click and the dedicated button. Delete them via right click > delete, right click > edit > delete, and drag and drop. Bookmarks are now saved and loaded to and from localStorage. Bookmark icon URLs are sent to our server, downloaded, encoded as base64, and sent back, and that is saved to the localStorag, so the user isn't downloading the images again, which in turn, means the images are loaded instantly, rather than waiting for them from a remote server.

Drag and drop for editing is problematic. It should work. but doesn't. No idea why.
Completed:
[x] Enable bookmark deletion via right click
[x] Begin work on saving bookmarks to ~~cookies~~ localStorage
Still to do:
[ ] Enable bookmark edit via drag and drop
[ ] Verify and santize all text inputs. Bookmark names need to be limited to certain characters, while also allowing characters from other languages. URLs need to be verified as real. The encoding proccess server-side needs to verify that the URL given is an image.
[ ] Begin work on the Steam Web API
[ ] Begin work on displaying links to wikis, videos and articles based on which game a given user is playing.
[ ] Research how to get HTTPS up and running, so that we may offer our website offline via Service Workers.
[ ] Find a name that is unused for website domain registration. If need be, update all mentions of WikiFinder to that name.
[ ] Once a suitable name is found, start work on integrating with Travis CI.

The README.MD now has a link to Liberapay, to support future development and maintenance.